+++ js/Button.js
@@ -21,11 +21,15 @@ function Button(options)
  }
  if ( options.fa ) { this.setIcon(options.fa); }
  if ( options.styles ) { for ( k in options.styles ) { if ( options.styles.hasOwnProperty(k) ) { this._node.style[k] = options.styles[k]; } } }
- if ( options.callback ) { this._node.onclick = (function(btn, cb) { return function() { cb.call(btn); }; })(this, options.callback); }
+ if ( options.callback ) { this._node.onclick = (function(btn, cb) { return function(evt) { if ( btn.isEnabled() ) { cb.call(btn, evt); } }; })(this, options.callback); }
  if ( options.small ) { this._node.classList.add("small"); }
  if ( options.listener ) { API.EVT.on(options.listener.method, options.listener.callback, options.listener.context || this); }
  if ( options.hide ) { this.hide(); }
  if ( options.active ) { this.setActive(options.active); }
+ if ( options.title )  { this._node.title = options.title; }
+ this.setEnabled(true);
+ if ( options.disabled ) { this.setDisabled(options.disabled); }
+ if ( options.enabled ) { this.setEnabled(options.enabled); }
  return this;
 }
 
@@ -40,6 +44,10 @@ Button.prototype =
  "show" : function() { this._node.style.display = ""; },
  "hide" : function() { this._node.style.display = "none"; },
  "getId" : function() { return this._id; },
+ "setDisabled" : function(state) { this._node.classList[state ? "add" : "remove"]("disabled"); },
+ "isDisabled" : function() { return this._node.classList.contains("disabled"); },
+ "setEnabled" : function(state) { this.setDisabled(!state); },
+ "isEnabled" : function() { return !this.isDisabled(); },
  "setIcon" : function(icon)
  {
   this.removeIcon();
@@ -112,4 +120,71 @@ ToggleSwitch.prototype =
  "show" : function() { this._node.style.display = ""; this._label.style.display = ""; },
  "hide" : function() { this._node.style.display = "none"; this._label.style.display = "none"; },
  "setDisabled" : function(state) { this._input.disabled = state; }
+};
+
+function Dropdown(options)
+{
+ var opt = {},
+     map = ["renderTo", "id", "noHover", "className", "content", "fa", "styles", "small", "listener", "hide", "active", "title"];
+ map.forEach(function(item) { if ( options[item] ) { opt[item] = options[item]; } });
+ opt.callback = this.showItems();
+ this.btn = new Button(opt);
+ this.items = options.items;
+}
+
+Dropdown.listenerDocument = false;
+
+Dropdown.prototype =
+{
+ "showItems" : function()
+ {
+  var that = this;
+  return function(evt)
+  {
+   var bounding = that.btn._node.getBoundingClientRect(),
+       ul = document.getElementById("dropdown-items");
+   if ( ul && ul.parentNode ) { ul.parentNode.removeChild(ul); }
+   ul = document.body.appendChild(document.createElement("ul"));
+   ul.id = "dropdown-items";
+   that.items.forEach(function(item)
+                      {
+                       var span,
+                           li = this.appendChild(document.createElement("li"));
+                       li.className = "dropdown-item";
+                       if ( item.fa )
+                       {
+                        span = li.appendChild(document.createElement("i"));
+                        li.classList.add("image");
+                        span.className = "fa fa-{0}".format(item.fa);
+                       }
+                       if ( item.content )
+                       {
+                        span = li.appendChild(document.createElement("span"));
+                        li.classList.add("content");
+                        span.innerHTML = item.content;
+                       }
+                       li.dataset.itemId = item.id;
+                      }, ul);
+   ul.style.top = (bounding.top + bounding.height) + "px";
+   ul.style.left = bounding.left + "px";
+   evt.stopPropagation();
+   if ( !Dropdown.listenerDocument )
+   {
+    document.body.addEventListener("click", function(evt)
+                                            {
+                                             var item,
+                                                 target = API.EVT.getParentTarget(evt, "li", "dropdown-item");
+                                                 ul = document.getElementById("dropdown-items");
+                                             if ( ul && ul.parentNode ) { ul.parentNode.removeChild(ul); }
+                                             if ( target )
+                                             {
+                                              item = that.items.find(function(item) { return item.id == target.dataset.itemId; });
+                                              if ( item && item.callback ) { item.callback(); }
+                                              return false;
+                                             }
+                                            }, false);
+    Dropdown.listenerDocument = true;
+   }
+  };
+ }
 };