+++ js/MODAL.js
@@ -1,9 +1,10 @@
 /* global API, Button */
 var MODAL =
 {
+ "buttons" : [],
  "show" : function(options)
  {
-  var wrapper, header, body, footer, closer, i;
+  var wrapper, header, actions, body, footer, closer, i;
   API.DOM.addLinkElement("css/modal.css");
   MODAL.hide();
   MODAL._node = document.body.appendChild(document.createElement("div"));
@@ -13,20 +14,32 @@ var MODAL =
   header = wrapper.appendChild(document.createElement("h1"));
   header.id = "modalHeader";
   closer = header.appendChild(document.createElement("span"));
-  closer.className = "fa fa-window-close";
+  closer.className = "fa fa-times";
   closer.onclick = MODAL.hide;
   header.appendChild(document.createTextNode(options.title || " "));
+  if ( options.actions )
+  {
+   actions = wrapper.appendChild(document.createElement("h2"));
+   actions.id = "modalAction";
+   options.actions.forEach(function(action) { this.appendChild(action); }, actions);
+   wrapper.classList.add("withActionBar");
+  }
   body = wrapper.appendChild(document.createElement("div"));
   body.id = "modalBody";
-  if ( options.body ) { body.innerHTML = options.body; }
+  if ( options.body )
+  {
+   if ( typeof options.body == "string" ) { body.innerHTML = options.body; }
+   else { body.appendChild(options.body); }
+  }
   footer = wrapper.appendChild(document.createElement("div"));
   footer.id = "modalFooter";
+  MODAL.buttons = [];
   if ( options.buttons )
   {
    for ( i = 0; i < options.buttons.length; i++ )
    {
     options.buttons[i].renderTo = footer;
-    new Button(options.buttons[i]);
+    MODAL.buttons.push(new Button(options.buttons[i]));
    }
   }
  },